requirements.txt
litellm>=1.83.0
python-dotenv==1.1.0
tiktoken==0.9.0
numpy==2.2.4
faiss-cpu==1.11.0
sentence-transformers==4.1.0
requests>=2.31.0
